separate useAgentForm
