+++ src/stores/agents/agentSlice.ts
@@ -195,9 +195,10 @@ const agentSlice = createSlice({
     hideSnackbar: (state) => ({
       ...state,
       snackbar: {
+        ...state.snackbar,
         open: false,
-        message: '',
-        severity: 'info',
+        // message: '',
+        // severity: 'info',
       },
     }),
   },
